requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.2.34
pandas>=1.1.4
seaborn>=0.11.0
dill==0.3.9


# Flask
# matplotlib
# torch
# gunicorn
# ultralytics